+++ src/cfg.cpp
@@ -128,6 +128,7 @@ std::vector<std::string> disable_animated_textures_on_field;
 long ff7_fps_limiter;
 bool ff7_footsteps;
 bool ff7_field_center;
+bool ff7_field_autosize_text_box;
 bool use_sdl_gamepad;
 bool ff7_japanese_text;
 bool enable_analogue_controls;
@@ -162,7 +163,6 @@ long display_index;
 long ff8_high_res_font;
 long hardware_video_decoding;
 std::vector<std::string> ff8_disable_remastered_hd_textures;
-bool ff7_field_autosize_text_box;
 
 std::vector<std::string> get_string_or_array_of_strings(const toml::node_view<toml::node> &node)
 {
@@ -299,6 +299,7 @@ void read_cfg()
 	ff7_fps_limiter = config["ff7_fps_limiter"].value_or(FPS_LIMITER_DEFAULT);
 	ff7_footsteps = config["ff7_footsteps"].value_or(false);
 	ff7_field_center = config["ff7_field_center"].value_or(true);
+	ff7_field_autosize_text_box = config["ff7_field_autosize_text_box"].value_or(true);
 	ff7_multibyte_font = config["ff7_multibyte_font"].value_or(false);
 	use_sdl_gamepad = config["use_sdl_gamepad"].value_or(false);
 	ff7_japanese_text = config["ff7_japanese_text"].value_or(false);
@@ -336,7 +337,6 @@ void read_cfg()
 	ff8_high_res_font = config["ff8_high_res_font"].value_or(-1);
 	hardware_video_decoding = config["hardware_video_decoding"].value_or(HWVA_NONE);
 	ff8_disable_remastered_hd_textures = get_string_or_array_of_strings(config["ff8_disable_remastered_hd_textures"]);
-	ff7_field_autosize_text_box = config["ff7_field_autosize_text_box"].value_or(true);
 
 	// Windows x or y size can't be less then 0
 	if (window_size_x < 0) window_size_x = 0;