+++ src/cfg.h
@@ -150,6 +150,7 @@ extern std::vector<std::string> disable_animated_textures_on_field;
 extern long ff7_fps_limiter;
 extern bool ff7_footsteps;
 extern bool ff7_field_center;
+extern bool ff7_field_autosize_text_box;
 extern bool use_sdl_gamepad;
 extern bool ff7_japanese_text;
 extern bool enable_analogue_controls;
@@ -185,7 +186,6 @@ extern long display_index;
 extern long ff8_high_res_font;
 extern long hardware_video_decoding;
 extern std::vector<std::string> ff8_disable_remastered_hd_textures;
-extern bool ff7_field_autosize_text_box;
 
 void read_cfg();
 bool is_remastered_hd_textures_disabled(const std::string &module);